pyproject.toml
[project]
name = "summerjob-matcher"
version = "0.1.0"
description = "Summer Job Matcher Streamlit App"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "numpy",
    "requests",
    "beautifulsoup4",
    "python-dotenv",
    "google-generativeai"
]

[tool.uv]
dev-dependencies = []
